requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
darkdetect==0.8.0
DateTime==5.5
distro==1.9.0
edge-tts==7.2.1
emoji==2.14.1
Flask==3.1.2
frozenlist==1.7.0
google-ai-generativelanguage==0.6.6
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyper-bot==0.78.2
idna==3.10
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multidict==6.6.4
numpy==2.3.2
openai==1.102.0
paramiko==4.0.0
pillow==11.3.0
playwright==1.54.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PyGithub==2.7.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
PySide6==6.9.2
PySide6-Fluent-Widgets==1.8.7
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
PySideSix-Frameless-Window==0.7.3
pytz==2025.2
pywin32==311
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
shiboken6==6.9.2
sniffio==1.3.1
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
winrt-runtime==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Media.Control==3.2.1
yarl==1.20.1
zope.interface==7.2